加入AgentWeb

+++ app/src/main/java/com/atlands/assistant/db/DBManager.java
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Environment;
-import android.support.annotation.Nullable;
 import android.util.Log;
 
 import java.io.File;
@@ -14,35 +13,51 @@ import java.io.IOException;
 import java.io.InputStream;
 
 public class DBManager {
-    private final int BUFFER_SIZE = 10000;
-    public static final String DB_NAME = "assistant"; //保存的数据库文件名
-    public static final String PACKAGE_NAME = "com.atlands.assistant";//工程包名
-    public static final String DB_PATH = "/data"
-            + Environment.getDataDirectory().getAbsolutePath() + "/"
-            + PACKAGE_NAME+"/databases";  //在手机里存放数据库的位置
+
+    private static final String DB_NAME = "assistant.db"; //保存的数据库文件名
+    private static final String PACKAGE_NAME = "com.atlands.assistant";
+    @SuppressLint("SdCardPath")
+    private static final String DB_PATH = "/data/data/com.atlands.assistant/" + "databases/";  //在手机里存放数据库的位置
+
+    private SQLiteDatabase database;
     private Context context;
 
     public DBManager(Context context) {
         this.context = context;
     }
 
-    public void copyDatabase() {
-        String dbfile=DB_PATH + "/" + DB_NAME ;
+    public void openDatabase() {
+        this.database = this.openDatabase(DB_PATH + "/" + DB_NAME);
+    }
+
+    private SQLiteDatabase openDatabase(String dbfile) {
         try {
-            //执行数据库导入
-            InputStream is = this.context.getResources().getAssets().open("assistant"); //欲导入的数据库
-            FileOutputStream fos = new FileOutputStream(dbfile);
-            byte[] buffer = new byte[BUFFER_SIZE];
-            int count = 0;
-            while ((count = is.read(buffer)) > 0) {
-                fos.write(buffer, 0, count);
+            if (!(new File(dbfile).exists())) {     //判断数据库文件是否存在，若不存在则执行导入，否则直接打开数据库
+                InputStream is = this.context.getResources().getAssets().open(DB_NAME);//欲导入的数据库
+                FileOutputStream fos = new FileOutputStream(dbfile);
+                int BUFFER_SIZE = 400000;
+                byte[] buffer = new byte[BUFFER_SIZE];
+                int count = 0;
+                while ((count = is.read(buffer)) > 0) {
+                    fos.write(buffer, 0, count);
+                }
+                fos.close();
+                is.close();
             }
-            fos.close();//关闭输出流
-            is.close();//关闭输入流
+            SQLiteDatabase db = SQLiteDatabase.openOrCreateDatabase(dbfile,
+                    null);
+            return db;
         } catch (FileNotFoundException e) {
+            Log.e("Database", "File not found");
             e.printStackTrace();
         } catch (IOException e) {
+            Log.e("Database", "IO exception");
             e.printStackTrace();
         }
+        return null;
+    }
+
+    public void closeDatabase() {
+        this.database.close();
     }
 }
